+++ package.json
@@ -21,6 +21,8 @@
   "devDependencies": {
     "@babel/core": "^7.12.9",
     "@expo/vector-icons": "^13.0.0",
+    "@react-navigation/native": "^6.1.2",
+    "@react-navigation/stack": "^6.3.11",
     "@types/react": "~17.0.21",
     "@types/react-native": "~0.66.13",
     "typescript": "~4.3.5"
